[CS-2669]: Change android bundle id
