+++ android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "me.rainbow",
+    package = "com.cardstack.cardpay",
 )
 
 android_resource(
     name = "res",
-    package = "me.rainbow",
+    package = "com.cardstack.cardpay",
     res = "src/main/res",
 )
