+++ android/app/src/androidTest/java/com/cardstack/cardpay/DetoxTest.java
@@ -1,4 +1,4 @@
-package me.rainbow;
+package com.cardstack.cardpay;
 
 
 import com.wix.detox.Detox;
@@ -25,7 +25,7 @@ public class DetoxTest {
         DetoxConfig detoxConfig = new DetoxConfig();
         detoxConfig.idlePolicyConfig.masterTimeoutSec = 90;
         detoxConfig.idlePolicyConfig.idleResourceTimeoutSec = 60;
-        detoxConfig.rnContextLoadTimeoutSec = (me.rainbow.BuildConfig.DEBUG ? 180 : 60);
+        detoxConfig.rnContextLoadTimeoutSec = (com.cardstack.cardpay.BuildConfig.DEBUG ? 180 : 60);
 
         Detox.runTests(mActivityRule, detoxConfig);
     }
