+++ android/app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <manifest 
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="me.rainbow">
+    package="com.cardstack.cardpay">
 
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.CAMERA" />
