+++ android/app/src/main/java/com/cardstack/cardpay/NativeModules/RNBip39/RNBip39Module.java
@@ -1,4 +1,4 @@
-package me.rainbow.NativeModules.RNBip39;
+package com.cardstack.cardpay.NativeModules.RNBip39;
 
 import com.facebook.react.bridge.Promise;
 import com.facebook.react.bridge.ReactContextBaseJavaModule;
