+++ android/app/src/main/java/com/cardstack/cardpay/NativeModules/RNBip39/RNBip39Package.java
@@ -1,4 +1,4 @@
-package me.rainbow.NativeModules.RNBip39;
+package com.cardstack.cardpay.NativeModules.RNBip39;
 
 import com.facebook.react.ReactPackage;
 import com.facebook.react.bridge.NativeModule;
