+++ android/app/src/main/java/com/cardstack/cardpay/NativeModules/RNZoomableButton/RNZoomableButtonPackage.java
@@ -1,22 +1,25 @@
-package me.rainbow.NativeModules.RNReview;
+package com.cardstack.cardpay.NativeModules.RNZoomableButton;
 
 import com.facebook.react.ReactPackage;
 import com.facebook.react.bridge.NativeModule;
 import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.uimanager.ViewManager;
+import com.swmansion.gesturehandler.react.RNZoomableButtonManager;
 
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 
-public class RNReviewPackage implements ReactPackage {
+public class RNZoomableButtonPackage implements ReactPackage {
+
     @Override
     public List<NativeModule> createNativeModules(ReactApplicationContext reactContext) {
-        return Arrays.asList(new RNReviewModule());
+        return Collections.emptyList();
     }
 
     @Override
     public List<ViewManager> createViewManagers(ReactApplicationContext reactContext) {
-        return Collections.emptyList();
+        return Arrays.<ViewManager>asList(
+                new RNZoomableButtonManager());
     }
 }
